// src/app/sdk/model/modelNarrativeDefinition.ts
/**
 * Robolucha API
 * Robolucha API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ModelNarrativeDefinition {
    event?: string;
    gameDefinition?: number;
    id?: number;
    mediaID?: string;
    sortOrder?: string;
    text?: string;
    type?: string;
}